pyproject.toml
[project]
name = "mcp-sse"
version = "0.1.0"
description = "SSE-based MCP Server and Client"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.45.1",
    "openai",
    "langchain",
    "langchain_groq",
    "argparse>=1.4.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.2.1",
    "python-dotenv>=1.0.1",
    "starlette>=0.45.3",
    "langgraph>=0.3.5",
]
